requirements.txt
# B&R Capital Dashboard Requirements
# Install with: pip install -r requirements.txt

# Core data processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.2
pyxlsb>=1.0.10  # For reading .xlsb files

# Database
psycopg2-binary>=2.9.7

# Web framework
streamlit>=1.28.0

# Visualization
plotly>=5.17.0
folium>=0.14.0
streamlit-folium>=0.15.0

# SharePoint/Azure
msal>=1.24.0  # Microsoft Authentication Library
requests>=2.31.0

# File monitoring
watchdog>=3.0.0

# Logging
structlog>=23.1.0

# Testing
pytest>=7.4.0

# Code formatting
black>=23.7.0

# Additional utilities
python-dotenv>=1.0.0  # For environment variables
xlwings>=0.30.0  # Alternative for Excel handling